@@ -15,8 +15,8 @@
         "source": "https://github.com/convenia/revisionable"
     },
     "require": {
-        "php": ">=5.3.0",
-        "illuminate/support": "~4.0|~5.0|~5.1|~5.2|~5.3"
+        "php": ">=5.6",
+        "illuminate/support": "~5.2|~5.3"
     },
     "autoload": {
         "classmap": [
