@@ -1,4 +1,8 @@
-<?php namespace Convenia\Revisionable;
+<?php
+
+namespace Convenia\Revisionable;
+
+use Illuminate\Database\Eloquent\Model as Eloquent;
 
 /*
  * This file is part of the Revisionable package by Venture Craft
@@ -8,65 +12,51 @@
  */
 
 /**
- * Class RevisionableTrait
- * @package Convenia\Revisionable
+ * Class Revisionable.
  */
-trait RevisionableTrait
+class Revisionable extends Eloquent
 {
     /**
-     * @var array
+     * @var
      */
-    private $originalData = array();
+    private $originalData;
+
+    /**
+     * @var
+     */
+    private $updatedData;
+
+    /**
+     * @var
+     */
+    private $updating;
 
     /**
      * @var array
      */
-    private $updatedData = array();
-
-    /**
-     * @var boolean
-     */
-    private $updating = false;
+    private $dontKeep = [];
 
     /**
      * @var array
      */
-    private $dontKeep = array();
+    private $doKeep = [];
 
     /**
-     * @var array
-     */
-    private $doKeep = array();
-
-    /**
-     * Keeps the list of values that have been updated
+     * Keeps the list of values that have been updated.
      *
      * @var array
      */
-    protected $dirtyData = array();
-
-    /**
-     * Ensure that the bootRevisionableTrait is called only
-     * if the current installation is a laravel 4 installation
-     * Laravel 5 will call bootRevisionableTrait() automatically
-     */
-    public static function boot()
-    {
-        parent::boot();
-
-        if (!method_exists(get_called_class(), 'bootTraits')) {
-            static::bootRevisionableTrait();
-        }
-    }
+    protected $dirtyData = [];
 
     /**
      * Create the event listeners for the saving and saved events
      * This lets us save revisions whenever a save is made, no matter the
      * http method.
-     *
      */
-    public static function bootRevisionableTrait()
+    public static function boot()
     {
+        parent::boot();
+
         static::saving(function ($model) {
             $model->preSave();
         });
@@ -75,7 +65,7 @@ trait RevisionableTrait
             $model->postSave();
         });
 
-        static::created(function($model){
+        static::created(function ($model) {
             $model->postCreate();
         });
 
@@ -94,26 +84,13 @@ trait RevisionableTrait
     }
 
     /**
-     * Generates a list of the last $limit revisions made to any objects of the class it is being called from.
+     * Invoked before a model is saved. Return false to abort the operation.
      *
-     * @param int $limit
-     * @param string $order
-     * @return mixed
+     * @return bool
      */
-    public static function classRevisionHistory($limit = 100, $order = 'desc')
-    {
-        return \Convenia\Revisionable\Revision::where('revisionable_type', get_called_class())
-            ->orderBy('updated_at', $order)->limit($limit)->get();
-    }
-
-    /**
-    * Invoked before a model is saved. Return false to abort the operation.
-    *
-    * @return bool
-    */
     public function preSave()
     {
-        if (!isset($this->revisionEnabled) || $this->revisionEnabled) {
+        if (! isset($this->revisionEnabled) || $this->revisionEnabled) {
             // if there's no revisionEnabled. Or if there is, if it's true
 
             $this->originalData = $this->original;
@@ -122,10 +99,9 @@ trait RevisionableTrait
             // we can only safely compare basic items,
             // so for now we drop any object based items, like DateTime
             foreach ($this->updatedData as $key => $val) {
-                if (gettype($val) == 'object' && !method_exists($val, '__toString')) {
+                if (gettype($val) == 'object' && ! method_exists($val, '__toString')) {
                     unset($this->originalData[$key]);
                     unset($this->updatedData[$key]);
-                    array_push($this->dontKeep, $key);
                 }
             }
 
@@ -147,7 +123,6 @@ trait RevisionableTrait
         }
     }
 
-
     /**
      * Called after a model is successfully saved.
      *
@@ -155,151 +130,122 @@ trait RevisionableTrait
      */
     public function postSave()
     {
-        if (isset($this->historyLimit) && $this->revisionHistory()->count() >= $this->historyLimit) {
-            $LimitReached = true;
-        } else {
-            $LimitReached = false;
-        }
-        if (isset($this->revisionCleanup)){
-            $RevisionCleanup=$this->revisionCleanup;
-        }else{
-            $RevisionCleanup=false;
-        }
 
         // check if the model already exists
-        if (((!isset($this->revisionEnabled) || $this->revisionEnabled) && $this->updating) && (!$LimitReached || $RevisionCleanup)) {
+        if ((! isset($this->revisionEnabled) || $this->revisionEnabled) && $this->updating) {
             // if it does, it means we're updating
 
             $changes_to_record = $this->changedRevisionableFields();
 
-            $revisions = array();
+            $revisions = [];
 
             foreach ($changes_to_record as $key => $change) {
-                $revisions[] = array(
-                    'revisionable_type' => $this->getMorphClass(),
-                    'revisionable_id' => $this->getKey(),
-                    'key' => $key,
-                    'owner_id' => $this->getOwnerId(),
-                    'old_value' => array_get($this->originalData, $key),
-                    'new_value' => $this->updatedData[$key],
-                    'user_id' => $this->getSystemUserId(),
-                    'created_at' => new \DateTime(),
-                    'updated_at' => new \DateTime(),
-                );
+                $revisions[] = [
+                    'revisionable_type'     => $this->getMorphClass(),
+                    'revisionable_id'       => $this->getKey(),
+                    'key'                   => $key,
+                    'old_value'             => array_get($this->originalData, $key),
+                    'new_value'             => $this->updatedData[$key],
+                    'user_id'               => $this->getSystemUserId(),
+                    'created_at'            => new \DateTime(),
+                    'updated_at'            => new \DateTime(),
+                ];
             }
 
             if (count($revisions) > 0) {
-                if($LimitReached && $RevisionCleanup){
-                    $toDelete = $this->revisionHistory()->orderBy('id','asc')->limit(count($revisions))->get();
-                    foreach($toDelete as $delete){
-                        $delete->delete();
-                    }
-                }
                 $revision = new Revision;
                 \DB::table($revision->getTable())->insert($revisions);
-                \Event::fire('revisionable.saved', array('model' => $this, 'revisions' => $revisions));
             }
         }
     }
 
     /**
-    * Called after record successfully created
-    */
+     * Called after record successfully created.
+     */
     public function postCreate()
     {
 
         // Check if we should store creations in our revision history
         // Set this value to true in your model if you want to
-        if(empty($this->revisionCreationsEnabled))
-        {
+        if (empty($this->revisionCreationsEnabled)) {
             // We should not store creations.
             return false;
         }
 
-        if ((!isset($this->revisionEnabled) || $this->revisionEnabled))
-        {
-
-            $revisions[] = array(
+        if ((! isset($this->revisionEnabled) || $this->revisionEnabled)) {
+            $revisions[] = [
                 'revisionable_type' => $this->getMorphClass(),
                 'revisionable_id' => $this->getKey(),
                 'key' => self::CREATED_AT,
                 'old_value' => null,
                 'new_value' => $this->{self::CREATED_AT},
-                'owner_id' => $this->getOwnerId(),
                 'user_id' => $this->getSystemUserId(),
                 'created_at' => new \DateTime(),
                 'updated_at' => new \DateTime(),
-            );
+            ];
 
             $revision = new Revision;
             \DB::table($revision->getTable())->insert($revisions);
-            \Event::fire('revisionable.created', array('model' => $this, 'revisions' => $revisions));
         }
-
     }
 
     /**
-     * If softdeletes are enabled, store the deleted time
+     * If softdeletes are enabled, store the deleted time.
      */
     public function postDelete()
     {
-        if ((!isset($this->revisionEnabled) || $this->revisionEnabled)
+        if ((! isset($this->revisionEnabled) || $this->revisionEnabled)
             && $this->isSoftDelete()
-            && $this->isRevisionable($this->getDeletedAtColumn())
-        ) {
-            $revisions[] = array(
+            && $this->isRevisionable($this->getDeletedAtColumn())) {
+            $revisions[] = [
                 'revisionable_type' => $this->getMorphClass(),
                 'revisionable_id' => $this->getKey(),
                 'key' => $this->getDeletedAtColumn(),
                 'old_value' => null,
-                'owner_id' => $this->getOwnerId(),
                 'new_value' => $this->{$this->getDeletedAtColumn()},
                 'user_id' => $this->getSystemUserId(),
                 'created_at' => new \DateTime(),
                 'updated_at' => new \DateTime(),
-            );
+            ];
             $revision = new \Convenia\Revisionable\Revision;
             \DB::table($revision->getTable())->insert($revisions);
-            \Event::fire('revisionable.deleted', array('model' => $this, 'revisions' => $revisions));
         }
     }
 
     /**
      * Attempt to find the user id of the currently logged in user
-     * Supports Cartalyst Sentry/Sentinel based authentication, as well as stock Auth
+     * Supports Cartalyst Sentry/Sentinel based authentication, as well as stock Auth.
      **/
-    public function getSystemUserId()
+    private function getSystemUserId()
     {
         try {
-            if (class_exists($class = '\SleepingOwl\AdminAuth\Facades\AdminAuth')
-                || class_exists($class = '\Cartalyst\Sentry\Facades\Laravel\Sentry')
-                || class_exists($class = '\Cartalyst\Sentinel\Laravel\Facades\Sentinel')
-            ) {
+            if (class_exists($class = '\Cartalyst\Sentry\Facades\Laravel\Sentry')
+                    || class_exists($class = '\Cartalyst\Sentinel\Laravel\Facades\Sentinel')) {
                 return ($class::check()) ? $class::getUser()->id : null;
             } elseif (\Auth::check()) {
                 return \Auth::user()->getAuthIdentifier();
             }
         } catch (\Exception $e) {
-            return null;
+            return;
         }
 
-        return null;
+        return;
     }
 
     /**
      * Get all of the changes that have been made, that are also supposed
-     * to have their changes recorded
+     * to have their changes recorded.
      *
      * @return array fields with new data, that should be recorded
      */
     private function changedRevisionableFields()
     {
-        $changes_to_record = array();
+        $changes_to_record = [];
         foreach ($this->dirtyData as $key => $value) {
             // check that the field is revisionable, and double check
             // that it's actually new data in case dirty is, well, clean
-            if ($this->isRevisionable($key) && !is_array($value)) {
-                if (!isset($this->originalData[$key]) || $this->originalData[$key] != $this->updatedData[$key]) {
+            if ($this->isRevisionable($key) && ! is_array($value)) {
+                if (! isset($this->originalData[$key]) || $this->originalData[$key] != $this->updatedData[$key]) {
                     $changes_to_record[$key] = $value;
                 }
             } else {
@@ -314,7 +260,7 @@ trait RevisionableTrait
     }
 
     /**
-     * Check if this field should have a revision kept
+     * Check if this field should have a revision kept.
      *
      * @param string $key
      *
@@ -338,7 +284,7 @@ trait RevisionableTrait
     }
 
     /**
-     * Check if soft deletes are currently enabled on this model
+     * Check if soft deletes are currently enabled on this model.
      *
      * @return bool
      */
@@ -346,7 +292,7 @@ trait RevisionableTrait
     {
         // check flag variable used in laravel 4.2+
         if (isset($this->forceDeleting)) {
-            return !$this->forceDeleting;
+            return ! $this->forceDeleting;
         }
 
         // otherwise, look for flag used in older versions
@@ -417,7 +363,7 @@ trait RevisionableTrait
     /**
      * Disable a revisionable field temporarily
      * Need to do the adding to array longhanded, as there's a
-     * PHP bug https://bugs.php.net/bug.php?id=42030
+     * PHP bug https://bugs.php.net/bug.php?id=42030.
      *
      * @param mixed $field
      *
@@ -425,8 +371,8 @@ trait RevisionableTrait
      */
     public function disableRevisionField($field)
     {
-        if (!isset($this->dontKeepRevisionOf)) {
-            $this->dontKeepRevisionOf = array();
+        if (! isset($this->dontKeepRevisionOf)) {
+            $this->dontKeepRevisionOf = [];
         }
         if (is_array($field)) {
             foreach ($field as $one_field) {
@@ -439,19 +385,4 @@ trait RevisionableTrait
             unset($donts);
         }
     }
-
-    public function getOwnerId()
-    {
-        $field = env('REVISION_OWNER_FIELD', 'identity_id');
-
-        $model = app($this->getMorphClass())->find($this->getKey());
-
-        if (isset($model->{$field}) ) {
-            return $model->{$field};
-        }
-
-        return 0;
-    }
-
-
 }
